[profiling] centralize profile-dir helpers
Move run-target parsing and profile-dir resolution into a shared Levanter utility module, then use it from the TensorBoard helper and Marin ingest path. This keeps the run-id logic and remote mirroring behavior in one place while preserving the existing CLI surface.
